pyproject.toml
[tool.poetry]
name = "todolist-app"
version = "0.1.0"
description = "A command-line To-Do List application built with Python OOP"
authors = ["Ali Ahmadi <mr-ahmadi2004@outlook.com>"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/yourusername/todolist-app"
repository = "https://github.com/yourusername/todolist-app"
documentation = "https://github.com/yourusername/todolist-app/docs"
keywords = ["todo", "todolist", "cli", "task-manager"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

packages = [{include = "todolist_app"}]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.7.0"
flake8 = "^6.1.0"
mypy = "^1.5.0"
isort = "^5.12.0"

[tool.poetry.scripts]
todolist = "todolist_app.cli.main:main"
todo = "todolist_app.cli.main:main"

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | \.mypy_cache
  | \.pytest_cache
  | __pycache__
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--cov=todolist_app",
    "--cov-report=html",
    "--cov-report=term-missing",
]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"